requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
pyyaml>=6.0
joblib>=1.3
tabulate

# Imbalance handling and tuning
imbalanced-learn>=0.11
optuna>=3.0

# Boosting backends
xgboost>=1.7
lightgbm>=4.0
catboost>=1.2

# Explainability
shap>=0.43
lime>=0.2.0.1

# Reporting
matplotlib>=3.7
seaborn>=0.12
